Change to run registry as a non-root user
This will allow us to run the registry in a
kubernetes cluster whose pod security policy
forbids containers running processes as root.

The default cache directory is /var/lib/registry
but I couldn't get that to work with a non-root
user. I think something in the base image must
be interfering, when I try to set permissions on
that path, so I've switched to 
/var/lib/docker-cache instead.