Add liveness probe: kill pod if it gets too big
We're using the container filesystem, so as we
pull more and more images via the cache, the 
container will consume more and more of the disk
space of the worker node on which it's running.

I don't want to use external file storage (e.g. 
s3) because that would make the cache slower, and
the whole point is to speed things up.

This change adds a liveness probe which will fail
if the amount of disk space used in the container
exceeds 5GB. When the liveness probe fails, the
cluster will restart the pod, emptying the cache.